Breaking out metrics, current, results & optout vote
